Update heat_furnace.lua
The connections weren't working properly because the object was 2x2 and not 3x3. The coordinates 1,0, and -1 define a 3x3. If the object has even size (2,4,6) it's connections need to be at -0.5 -> 0.5

Also, I added connections for all spots. Because it's not a 3x3 mirroring / rotating might not work if the spots don't exist on all spots.